Rewrote page 10-postgresql-trigger-security.mdx
Ref:

"Triggers in PostgreSQL execute with the privileges of the user who created the trigger, not the user who initiated the operation that fired the trigger."

--> Look into the documentation from PostgreSQL 18: it's the role who queued the trigger event or in the words of your website it's the user who initiated the operation that fired the trigger not the creator.

The documentation says: Also, the trigger will always run as the role that queued the trigger event, unless the trigger function is marked as SECURITY DEFINER, in which case it will run as the function owner.

https://www.postgresql.org/docs/current/trigger-definition.htm

"Use SECURITY DEFINER with SET ROLE to implement least privilege principles"

Actually this also really puzzled me. The first mistake is that it's about functions and not trigger in the topic trigger security. Okay there are trigger functions. Secondly the way the SECURITY DEFINER is used with the SET ROLE doesn't even work this way in the audit example. Thirdly the use of SECURITY DEFINER should be avoided as it is a security risk itself and should be used on a minimal level with care.

https://www.red-gate.com/simple-talk/databases/postgresql/securing-postgresql-against-injection-misconfiguration-and-privilege-escalation/